requirements.txt
google-genai
pydantic>=2.0
pathspec
